requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.5
django-on-heroku==1.1.2
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
static3==0.7.0
whitenoise==5.2.0
<<<<<<< HEAD
=======
django-heroku==0.3.1
psycopg2==2.8.6
requests==2.25.1
sqlparse==0.4.1
virtualenv==20.2.1
zipp==3.4.0
>>>>>>> 28db349d9fc26775f91308aec53ddc052260599d
